Loosens the restrictions on filters of workflows.
Workflows previously allowed for filters, which typically looked something like
this:

```python
@calc('a')
def flt_a(a):
    "Ensures that parameter `a` is a read-only numpy array."
    a = np.array(a)
    a.flags.writeable = False
    return (a,)
...
workflow = plan(
    filter_a=flt_a,
    ...)
```

These filters required that (1) the calc function have exactly 1 input
and exactly 1 output, both of which were the same, and (2) that the
key for that calc in the plan begin with the string `'filter_'`. This
is clunky and doesn't allow easily for scenarios where multiple
parameters need to be co-filtered, such as when any two of three
parameters may be provided to the workflow (with the workflow logic
then filling in the missing value).

This commit removes these requirements and instead enforces the following
requirement: each parameter may optionally appear the outputs of 1 calc in
the plan. Note that in order to be considered a parameter, it must also appear
in the calc's inputs. This is essentially the one exception to the rule of plans
that there can be no dependency cycle. The resulting plandict uses the output
value of each parameter as opposed to the provided value when applicable.